requirements.txt
pymupdf            # Extracción de texto y metadatos desde PDFs (rápido y preciso)
pytesseract        # OCR para PDFs escaneados o imágenes con texto
Pillow             # Procesamiento básico de imágenes (soporte para OCR)
numpy              # Operaciones numéricas y manejo de arreglos
pandas             # Manejo de datasets (preguntas, respuestas, metadatos)
python-dotenv      # Cargar variables del archivo .env (rutas, claves, configs)
tqdm               # Barras de progreso para procesos largos (chunking, OCR, etc.)
# Fase 3
transformers
sentence-transformers
# Fase 4
faiss-cpu
# Fase 5
json
pathlib
# LLM
hf_xet
# UI
streamlit
torch
accelerate